Building a base for the frontend
Removed all previously used .jade files and its integration within the server. I never used Jade and it doesn't seem to help with the project. 
Because of the small scale of the project i have chosen to go with jquery as a front-end library and started with building a basic foundation. Already implemented some basic routes and a favicon.